third_party/blink/web_tests/flag-specific/highdpi/fast/events/touch/gesture/gesture-tap-frame-overlap-expected.txt

This is a testharness.js-based test.
FAIL gesture-tap-frame-overlap assert_equals: expected 1 but got 0
Harness: the test ran to completion.
